refactor(gba): 卡带库与 press 计划从会话状态机拆出（849→678 行）
GbaService 混装 ROM 增删查改 + 帧循环 + 按键计划 + SRAM/快照持久化 + 看门狗。把两块与
模拟器核心无关的职责拆出去：

- rom-library.ts：ROM 增删查改（只碰 store + OSS）。deleteRom 的会话态把关（加载中 /
  正被玩一律先拒、deleteInFlight 互斥）**留在 service**——那是会话状态不是库的事；库只
  负责删除机制本身。MAX_ROM_BYTES / toRomView 随之迁移，两个 importer 改为直接导入真实路径。
- press-plan.ts：buildPressPlan 与 press 领域上限（纯函数，不碰会话状态）+ EMPTY_HELD。

帧循环刻意不动：它与 core/plan/foreground/看门狗/SRAM 深度耦合，「帧推进权唯一归属帧循环」
正是这个服务的本体职责，抽出去只会把耦合变成一堆回调。纯重构，40 个 gba 测试全绿。
